+++ R/lp.R
@@ -6,6 +6,7 @@
 #' @param formula formula specifying the linear predictors.
 #' @param coefs vector of regression coefficients.
 #' @param data data frame containing the covariates used to construct the linear predictors.
+#' @param ... further arguments passed to other methods.
 #' @return a vector containing the linear predictors.
 #'
 #' @examples
@@ -25,13 +26,13 @@
 #' glimpse(simdata)
 #'
 
-lp <- function(formula, coefs, data){
+lp <- function(formula, coefs, data, ...){
   if(missing(data)){
     mf <- stats::model.frame(formula=formula)
   }else{
     mf <- stats::model.frame(formula=formula, data=data)
   }
-  X <- stats::model.matrix(mf)
+  X <- stats::model.matrix(mf, ...)
   lp <- as.numeric(X%*%coefs)
   return(lp)
 }