+++ R/raftreg.R
@@ -9,6 +9,8 @@
 #' @param beta vector of regression coefficients.
 #' @param dist an alternative way to specify the baseline survival distribution.
 #' @param package the name of the package where the assumed quantile function is implemented.
+#' @param lwr left-truncation time (default to 0 in the absence of left-truncation).
+#' @param upr right-truncation time (default to Inf in the absence of right-truncation).
 #' @param data data frame containing the covariates used to generate the survival times.
 #' @param ... further arguments passed to other methods.
 #' @return a numeric vector containing the generated random sample.
@@ -16,6 +18,7 @@
 #' @examples
 #' library(rsurv)
 #' library(dplyr)
+#' set.seed(123)
 #' n <-  1000
 #' simdata <- data.frame(
 #'   age = rnorm(n),
@@ -33,7 +36,7 @@
 #'   )
 #' glimpse(simdata)
 #'
-raftreg <- function(u, formula, baseline, beta, dist = NULL, package = NULL, data, ...){
+raftreg <- function(u, formula, baseline, beta, dist = NULL, package = NULL, lwr = 0, upr = Inf, data, ...){
   call <- match.call()
   if(!is.null(dist)){
     baseline <- dist
@@ -69,7 +72,7 @@ raftreg <- function(u, formula, baseline, beta, dist = NULL, package = NULL, dat
     lp <- lp + off
   }
 
-  time <- rEH(u, baseline = baseline, lp1 = lp, lp2 = -lp, package = package, ...)
+  time <- rEH(u, baseline = baseline, lp1 = lp, lp2 = -lp, package = package, lwr = lwr, upr = upr, ...)
   attributes(time) <- list(call = call, model.matrix=X, beta = beta)
   return(time)
 }