+++ tests/testthat/test1.R
@@ -5,7 +5,7 @@ n <-  1000
 simdata <- data.frame(
   age = rnorm(n),
   sex = sample(c("f", "m"), size = n, replace = TRUE)
-) |>
+) %>%
   mutate(
     aft = raftreg(runif(n), ~ age*sex, beta = c(1, 2, -0.5), dist = "lnorm", meanlog = 0, sdlog = 1),
     ph = rphreg(runif(n), ~ age*sex, beta = c(1, 2, -0.5), dist = "lnorm", meanlog = 0, sdlog = 1),